requirements.txt
anthropic>=0.40.0
streamlit>=1.29.0
python-dotenv>=1.0.0
PyYAML>=6.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
pytest>=8.0.0
